WordPressAuthenticator: Splitting Notifications Extension
